move conftest location, clean up test
